requirements.txt
pillow
pickleshare
numpy
tensorflow-cpu
streamlit
urllib3